refactor: repackage AuthHeaderResolver and PassportMapper
