+++ src/main/java/com/lxp/gateway/jwt/filter/JwtAuthenticationGatewayFilterFactory.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.lxp.common.infrastructure.exception.ErrorResponse;
 import com.lxp.gateway.global.constants.MDCConstants;
 import com.lxp.gateway.global.exception.GatewayErrorCode;
-import com.lxp.gateway.jwt.adapter.AuthHeaderResolver;
-import com.lxp.gateway.jwt.adapter.PassportMapper;
+import com.lxp.gateway.jwt.support.AuthHeaderResolver;
+import com.lxp.gateway.jwt.support.PassportMapper;
 import com.lxp.gateway.jwt.exception.InvalidTokenException;
 import com.lxp.gateway.jwt.policy.JwtPolicy;
 import com.lxp.gateway.jwt.policy.TokenRevocationPolicy;
